+++ tests/fixtures/create_fixtures.py
@@ -3,8 +3,8 @@ from pathlib import Path
 
 import numpy as np
 from bw_processing import INDICES_DTYPE, UNCERTAINTY_DTYPE, create_datapackage
-from fs.osfs import OSFS
-from fs.zipfs import ZipFS
+from bw_processing.io_helpers import generic_directory_filesystem
+from fsspec.implementations.zip import ZipFileSystem
 
 fixture_dir = Path(__file__).resolve().parent
 
@@ -38,7 +38,7 @@ def empty_biosphere():
     # Activity 1: The activity
 
     dp = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "empty_biosphere.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "empty_biosphere.zip", write=True),
     )
 
     data_array = np.array([1, 2, 3])
@@ -119,28 +119,29 @@ def _create_basic_fixture(fs, characterization=True, characterization_values=Tru
 
 
 def create_basic_fixture_zipfile():
-    _create_basic_fixture(ZipFS(str(fixture_dir / "basic_fixture.zip"), write=True))
+    _create_basic_fixture(ZipFileSystem(fixture_dir / "basic_fixture.zip", write=True))
 
 
 def create_missing_characterization():
     _create_basic_fixture(
-        ZipFS(str(fixture_dir / "missing_characterization.zip"), write=True), characterization=False
+        ZipFileSystem(fixture_dir / "missing_characterization.zip", write=True),
+        characterization=False,
     )
 
 
 def create_empty_characterization():
     _create_basic_fixture(
-        ZipFS(str(fixture_dir / "empty_characterization.zip"), write=True),
+        ZipFileSystem(fixture_dir / "empty_characterization.zip", write=True),
         characterization_values=False,
     )
 
 
 def create_basic_fixture_directory():
-    _create_basic_fixture(OSFS(str(fixture_dir / "basic_fixture"), create=True))
+    _create_basic_fixture(generic_directory_filesystem(dirpath=fixture_dir / "basic_fixture"))
 
 
 def create_svdm_fixtures():
-    dp = create_datapackage(fs=ZipFS(str(fixture_dir / "svdm.zip"), write=True), sequential=True)
+    dp = create_datapackage(fs=ZipFileSystem(fixture_dir / "svdm.zip", write=True), sequential=True)
 
     data_array = np.array([42])
     indices_array = np.array([(1, 1)], dtype=INDICES_DTYPE)
@@ -169,7 +170,7 @@ def create_svdm_fixtures():
 
     dp.finalize_serialization()
 
-    dp2 = create_datapackage(fs=ZipFS(str(fixture_dir / "svdm2.zip"), write=True))
+    dp2 = create_datapackage(fs=ZipFileSystem(fixture_dir / "svdm2.zip", write=True))
 
     data_array = np.array([88])
     indices_array = np.array([(2, 2)], dtype=INDICES_DTYPE)
@@ -187,7 +188,7 @@ def create_array_fixtures():
     # Products: 1, 2
     # Biosphere flows: 1
     dp = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "array_sequential.zip"), write=True), sequential=True
+        fs=ZipFileSystem(fixture_dir / "array_sequential.zip", write=True), sequential=True
     )
 
     data_array = np.array([1, 1, 0.5])
@@ -232,7 +233,7 @@ def create_mc_basic():
     # Activity 1
     # Activity 2
     dp = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "mc_basic.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "mc_basic.zip", write=True),
     )
 
     data_array = np.array([1, 1, 0.5])
@@ -302,7 +303,7 @@ def create_multilca_simple():
     #   N: 102  -0.5    1
 
     dp1 = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "multi_lca_simple_1.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "multi_lca_simple_1.zip", write=True),
     )
     data_array = np.array([0.2, 0.5, 1, 1, 0.5, 1])
     indices_array = np.array(
@@ -334,7 +335,7 @@ def create_multilca_simple():
     #   P: 104  1
 
     dp2 = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "multi_lca_simple_2.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "multi_lca_simple_2.zip", write=True),
     )
     data_array = np.array([0.1, 0.2, 0.1, 0.4, 1, 1])
     indices_array = np.array(
@@ -356,7 +357,7 @@ def create_multilca_simple():
     #   Q: 105  1
 
     dp3 = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "multi_lca_simple_3.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "multi_lca_simple_3.zip", write=True),
     )
     data_array = np.array([0.1, 0.2, 1])
     indices_array = np.array([(100, 6), (102, 6), (105, 6)], dtype=INDICES_DTYPE)
@@ -427,7 +428,7 @@ def create_multilca_simple():
     # 200-206   1
 
     dp4 = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "multi_lca_simple_4.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "multi_lca_simple_4.zip", write=True),
     )
     indices_array = np.array(
         [(200, 0), (201, 0), (202, 0), (203, 0), (204, 0), (205, 0), (206, 0)], dtype=INDICES_DTYPE
@@ -460,7 +461,7 @@ def create_multilca_simple():
     # 205   10
 
     dp5 = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "multi_lca_simple_5.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "multi_lca_simple_5.zip", write=True),
     )
     indices_array = np.array([(201, 0), (203, 0), (205, 0)], dtype=INDICES_DTYPE)
     distributions_array = np.array(
@@ -482,7 +483,7 @@ def create_multilca_simple():
     )
 
     dp6 = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "multi_lca_simple_normalization.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "multi_lca_simple_normalization.zip", write=True),
     )
     indices_array = np.array([(201, 0), (203, 0)], dtype=INDICES_DTYPE)
     distributions_array = np.array(
@@ -503,7 +504,7 @@ def create_multilca_simple():
     )
 
     dp7 = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "multi_lca_simple_weighting.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "multi_lca_simple_weighting.zip", write=True),
     )
     indices_array = np.array([(0, 0)], dtype=INDICES_DTYPE)
     distributions_array = np.array(
@@ -538,7 +539,7 @@ def create_mc_complete():
     # Activity 1
     # Activity 2
     dp = create_datapackage(
-        fs=ZipFS(str(fixture_dir / "mc_complete.zip"), write=True),
+        fs=ZipFileSystem(fixture_dir / "mc_complete.zip", write=True),
     )
 
     data_array = np.array([1, 2])